adds .kill to scramble
