+++ scramble/.cs50.yaml
