+++ scramble/words.txt
@@ -1,4 +1,3 @@
-A
 AA
 AAH
 AAHED
